fixed an issue with selecting members by keyboard
