+++ src/components/PeoplePicker/PeoplePickerItemProps.ts
@@ -1,10 +1,9 @@
 import {
-  IPersonaProps,
-  PersonaSize
+  IPersonaProps
 } from '../../Persona';
 import { PeoplePickerType } from './PeoplePicker.Props';
 import * as React from 'react';
-export interface IPeoplePickerItemProps extends React.Props<any>{
+export interface IPeoplePickerItemProps extends React.Props<any> {
   persona: IPersonaProps;
   onRemovePersona?: (index: number, personaProps: IPersonaProps) => void;
   onSelectPersona?: (personaProps: IPersonaProps) => void;
